+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 ### Fixed
 - Fixed an error where users were unable to setup mysql databases.
 - Fixed an error when creating a backup with `db backup` was not including the drop tables command.
+- Fixed an error when installing extensions would not reflect in the site container.
 
 ## 2.0.8 - 2021-05-18